requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
python-decouple
asyncpg
argon2-cffi
python-jose[cryptograpy]
python-multipart
pillow
transformers